Logisim Testing
Here is a test file made in Logisim. If someone could download it to see if it works on their version and report back if it does that would be great thx. Also preferable if you test it on either windows or linux as this was made on Mac to test cross platform.